(#1540) Column database ID passed from front end
Instead of column index

+++ WebApp/src/uk/ac/exeter/QuinCe/web/datasets/plotPage/PlotPage2Data.java
@@ -34,7 +34,7 @@ public abstract class PlotPage2Data {
   /**
    * The column for which values have been selected
    */
-  protected int selectedColumn = -1;
+  protected long selectedColumn = -1L;
 
   /**
    * The IDs of the selected rows
@@ -403,7 +403,7 @@ public abstract class PlotPage2Data {
    *
    * @return The selected column.
    */
-  public int getSelectedColumn() {
+  public long getSelectedColumn() {
     return selectedColumn;
   }
 
@@ -413,7 +413,7 @@ public abstract class PlotPage2Data {
    * @param selectedColumn
    *          The selected column.
    */
-  public void setSelectedColumn(int selectedColumn) {
+  public void setSelectedColumn(long selectedColumn) {
     this.selectedColumn = selectedColumn;
   }
